2015 day 19

+++ lib/libcollections/src/matrix.c
@@ -22,7 +22,7 @@ Matrix* matrix_new(unsigned int width,
     m->width = width;
     m->height = height;
     m->data = ec_malloc(sizeof(int*) * height);
-    for (int i = 0; i < height; ++i) {
+    for (size_t i = 0; i < height; ++i) {
         m->data[i] = ec_calloc(width, sizeof(int));
     }
     return m;
@@ -30,8 +30,8 @@ Matrix* matrix_new(unsigned int width,
 
 void matrix_copy(Matrix* src, Matrix* dest) {
     assert(dest->height == src->height && dest->width == src->width);
-    for (int i = 0; i < dest->height; ++i) {
-        for (int j = 0; j < dest->width; ++j) {
+    for (size_t i = 0; i < dest->height; ++i) {
+        for (size_t j = 0; j < dest->width; ++j) {
             dest->data[i][j] = src->data[i][j];
         }
     }
@@ -40,8 +40,8 @@ void matrix_copy(Matrix* src, Matrix* dest) {
 int matrix_sum(Matrix* m)
 {
     int sum = 0;
-    for (int i = 0; i < m->height; ++i) {
-        for (int j = 0; j < m->width; ++j) {
+    for (size_t i = 0; i < m->height; ++i) {
+        for (size_t j = 0; j < m->width; ++j) {
             sum += m->data[i][j];
         }
     }
@@ -50,14 +50,14 @@ int matrix_sum(Matrix* m)
 
 int matrix_neighbors8(Matrix* matrix, int *buf, uint32_t x, uint32_t y)
 {
-    int xmin = max(0, x-1);
-    int xmax = min(matrix->width-1, x+1);
-    int ymin = max(0, y-1);
-    int ymax = min(matrix->height-1, y+1);
+    size_t xmin = max(0, x-1);
+    size_t xmax = min(matrix->width-1, x+1);
+    size_t ymin = max(0, y-1);
+    size_t ymax = min(matrix->height-1, y+1);
 
     int index = 0;
-    for (int i = ymin; i <= ymax; ++i) {
-        for (int j = xmin; j <= xmax; ++j) {
+    for (size_t i = ymin; i <= ymax; ++i) {
+        for (size_t j = xmin; j <= xmax; ++j) {
             if (i == y && j == x)
                 continue;
             buf[index++] = matrix->data[i][j];
@@ -68,8 +68,8 @@ int matrix_neighbors8(Matrix* matrix, int *buf, uint32_t x, uint32_t y)
 
 void matrix_print(Matrix* m)
 {
-    for (int i = 0; i < m->height; ++i) {
-        for (int j = 0; j < m->width; ++j) {
+    for (size_t i = 0; i < m->height; ++i) {
+        for (size_t j = 0; j < m->width; ++j) {
             printf("%d", m->data[i][j]);
         }
         printf("\n");
@@ -78,7 +78,7 @@ void matrix_print(Matrix* m)
 
 void matrix_free(Matrix *matrix)
 {
-    for (int i = 0; i < matrix->height; ++i) {
+    for (size_t i = 0; i < matrix->height; ++i) {
         free(matrix->data[i]);
     }
     free(matrix->data);
